requirements.txt
# WhatsApp AI Agent Requirements
# Core dependencies for the WhatsApp AI agent

# Core Python libraries
# asyncio is part of Python standard library (Python 3.4+)
typing-extensions>=4.5.0

# Configuration management
PyYAML>=6.0
python-dotenv>=1.0.0

# AI/ML libraries
openai>=1.3.0
anthropic>=0.7.0  # Optional: for Anthropic Claude integration

# WhatsApp integration libraries
# Note: Choose one based on your preferred approach
# whatsapp-web.js integration (via subprocess)
# selenium>=4.15.0  # For WhatsApp Web automation
# requests>=2.31.0  # For API calls

# Async HTTP client
aiohttp>=3.8.0
httpx>=0.25.0

# Utilities
python-dateutil>=2.8.0
pytz>=2023.3

# Development dependencies (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.9.0
flake8>=6.1.0
mypy>=1.6.0

# Logging and monitoring
structlog>=23.2.0

# Database (for conversation history - optional)
# sqlalchemy>=2.0.0
# aiosqlite>=0.19.0  # For SQLite async support
# redis>=5.0.0  # For Redis-based session storage